requirements.txt
langchain
crewai
streamlit
openai
python-dotenv
pydantic
pandas
requests
chromadb
sentence-transformers
ortools
